Agent crash @ DBTableWalker::WalkCancel
Issue:
The sequence of events is as follows:
- Notify walk going for a peer
- BGP peer goes down
- Stop of notify walk issued
- some time later delete of path done.

In third step i.e. stop of route walk causes all route table walks to be stopped
and cancelled. However it does not reset the route walkid. Now when delete path
is scheduled it also tries cancelling walk as walk id was not removed. DBwalker
does not have any knowledge of walk as it was stopped and it crashes.

Solution:
Clean the walker id on explicit stop walk in AgentRouteWalker.

Change-Id: Ie5da98ae44d4b7887ea0e30f61c7520993d8eec7
Closes-bug:#1701123