Stale logical switch never gets deleted.
As described in bug, logical switch which is stale is not getting deleted even
after stale cleanup because of reference pending.
Also multicast remote entry is not withdrawn as there is no corresponding stale
state for same.
Another place which block LS deletion is vlan-port binding. This vlan-port has
been moved from old LS to some new LS. This update is seen from config but does
not get percolated to Tor. The reason for same is that vlan-port expects new LS
to be active which is not, as its pending on vxlan resource being used by old LS.
In turn Tor as well sees that there is a vlan-port pointing to old LS and does
not acknowledge delete request. System enters deadlock for the old and new LS.

Solution:
On stale timer expiration release references, send duplicate deletes for
multicast route and LS to TOR. Also delete and re-add the vlan-port binding as
soon as a change in LS in observed. Delete will ensure Tor also releases old LS
for this vlan-port.

Conflicts:
	src/ksync/ksync_entry.h
	src/vnsw/agent/oper/vm_interface.h

Change-Id: Ie551ce27d76dd5d6c1043351345803a07738132a
Closes-bug: #1702431