fix bug1697542; add unit test for MatchFilter()
The analytics API of user-defined log statistic does not returns blank, even if
UVE is working properly and user-defined logs are created.  The created
name-pattern mapping is succesfully stored in the database. The counters to
each mapping, however, would not increment on hit.  The counters are handled
inside UserDefinedCounters::MatchFilter.  It turns out that
UserDefinedCounters::MatchFilter inside DbHandler::MessageTableKeywordInsert is
never called because the latter would always return in previous if statement,
before calling the matching function. 

By relocating the if section, MatchFilter will be called, and the counters
would be correctly handled.  When a counter is greater than 0, the
corresponding name-pattern mapping would be written to database, and therefore
would be succesfully fetched on API query.

Closes-Bug: 1697542
Change-Id: I9db6c296f6cef94d30981837436c565e366280f1