Fix nodemgr reporting process state info on 16.04 container
On 16.04 docker we are using systemd inside the docker
as an init system to launch the required processes. However
sd_notify call does not work in the docker and hence
nodemgr does not receive unit change notifications from
systemd system dbus. Fix is to instead do a poll of the
unit status when running inside the docker.

Change-Id: Ie5de608dff74de5287bfa6b9f6e4661f9c30250c
Closes-Bug: #1694617
Closes-Bug: #1694849