[VNC OpenStack] Use first VMI of the first SI's VM as router gateway interface
Contrail model diverges strongly from the Neutron model regarding the
way router gateway is implemented. Contrail uses a service instance to
insert a SNAT service between private and public networks. And
instantiates two VM with VMIs on that networks.
That patch proposes to use the first right VMI of the first SI's VM as
Neutron port for a router's gateway interface. It also limits the
visibility of that router gateway interface to admin users as it's done
in Neutron [1].

[1] https://github.com/openstack/neutron/blob/master/neutron/db/l3_db.py#L350

Change-Id: Ic6141c6dce8e56f95c81c492bc2e59fadea67ced
Closes-Bug: #1509101