E2 code changes
This commit contains the following changes.

Before the code is walked through please take a look at the E2 data model
that explains the need for the data objects. Below is the link to
the spec

https://github.com/Juniper/contrail-controller/blob/master/specs/contrail-E2.md

Changes are over the following modules.

1. Schema:
 - Contains the p2p service schema
 - Added telemetry support objects to physical-router
 - Also misc attributes like snmp, lldp etc.
 - And indexes for keeping track of services, physical indexes allocation etc.
 - Unit test cases

2. Device manager
 - Changes to support the following p2p services
   - l2circuit (port, virtual port)
   - l2vpn (port, virtual port)
   - PWHT
   - Fabric service

   - Fetch network config
   - Provide service status for a given service.

   - Nokia(ALU) specific services
     - Currently supported for l2circuit and l2vpn

   - Unit test cases will be addressed in next commit

3. UVE objects

Two new UVE's are being created for E2, namely 'physical_router_config' and
'service_status'. One is for collecting the router config and the other for
collecting the service status.

Since this UVE's are static and triggered from E2 application, the schema
data objects has also been created.

Nokia/ALU code is not properly working with the new plugin architecture, will
fix it in next commit for E2 code to work

Change-Id: I3fe3f8e4042bf71db009b241511a447f6622dfa1
Partial-Bug: #1700162