rmmod of vrouter to wait, to upgrade compute.
Description: stopping supervisor-vrouter is needed to
remove vrouter module. But stopping supervisor-vrouter
might take some time, which is causing a failure in
rmmod of vrouter, as vrouter is still in use by
contrail-vrouter-agent. With -w option, rmmod will
wait till vrouter module is not in use before it is
removed.

Change-Id: I68a2343fccbfc57f2930f393023cf558f39ac476
Closes-Bug:1680653