Implement community-all and community-any match in routing policy
Add two new keywords community-all and community-any for community
matching in addition to the existing community keyword which matches
exactly one community value. The API server will ensure that only
one of these 3 keywords can be used in a given routing policy term.

Both of the new keywords accept a list of values, each of which can
be a fixed community or a UNIX-style extended regular expression.
An intentional side effect of the new implementation is that the
community keyword now accepts a single regular expression or fixed
community value.

The community-all keyword implies that the communities in the path
being evaluated must match all match values specified in the list.
Note that it's fine for the path to have one or more communities that
do not match anything in the list, as long as all the values in the
list in the match condition are matched.

The community-any keyword implies that the communities in the path
being evaluated must match at least one value specified in the list.

Change-Id: Id98cb8aaf42d8cc211ea2f6141a4641f28292e1b
Partial-Bug: 1699586