Do not generate vrf assign rules for book-end RIs
When one port of a service VM is used in multiple service instances, the VRF
assign rules generated for this interface cannot identify which of those VRF
to use for the traffic. This is because VRF assign rules currently don't have
src or dst VN information. With this change, we will not create VRF assign
rules for 'book-end' interfaces, i.e., the left interface of the first instance
and right interface in the last instance in the chain. This allows us to
support this case when such a shared interface is the first or last in the
service chain.

Change-Id: Ibe9882fa96d4279b67c9aea469445c358505d957
Closes-Bug: 1661134