Return empty list if obj_uuid or back_ref_id are passed as empty
If a floating ip is not attached to a port, in vnc_openstack code,
we might call virtual_machine_interfaces_list with obj_uuids set to
empty. In this case, current library code will not set that field
in the request body causing API server to return all VMIs. To avoid
this situation, we now return early with empty list if either obj_uuids
or back_ref_id is set as empty list. Old behavior will remain if any
of these arguments are not passed or are passed as None.

Change-Id: I715ef9c60461f24e7523de4522ede852574474f5
Closes-Bug: 1706260
(cherry picked from commit 235caed6916ff8c19d0786df5a73972cd35b8572)