Operations on objects created pre RBAC fail, on enabling RBAC.
Description: Certain objects created pre RBAC has owner ID not
same as the id of the project in which those objects are created.
This is because the project information passed was the service
tenant. This is causing failure of operations on those objects
after enbaling RBAC.
This script iterates through all objects in project and changes
it's owner field to the project id.
Help on how to use the script is available in the script.
This is a partial fix. Comeplete fix will be available once we
neutron handler code in API server.

Change-Id: I478c967746a525eab156a8836fc7580518a384a3
Partial-Bug: #1706218