RBAC - shared list not working for neutron/WebUI
Description: Multiple issues are seen here.
In the API server, project id was not in uuid form.
Sometimes a project may not have an associated domain.
This will make shared=true work on API server.
Neutron relies on API lib to post the requests, so made
change in library to accept information and is backward
compatible for other applications.
The behavior in neutron going to be as follows:
1) neutron net-list would list
   a) all local networks in that tenant.
   b) globally shared.
   c) tenant shared.
2) neutron net-list --shared True would list
   a) globally shared.
   b) tenant shared.
3) neutron net-list --shared False would list
   a) all local networks in that tenant.

Change-Id: Iba10854e4a123454356c78f9a86b2eda1a588377
Closes-Bug:1707175