Improved error reporting in VNC Config API
Juniper contrail provides a REST based API that can be called by
external clients such as UI, openstack horizon or openstack neutron.
In case of success, it returns the information as a JSON formatted
data. But in case of failure, it just returns an HTTP error code
along with a message in string format. It is desirable to provide
extra information about what failed so that for the human users,
it can be presented to them in a more meaningful way and for clients
to take appropriate action based on the information.
Revamped whole error reporting for VNC Config API restructuring the
error classes and providing more details from the places where being
raised. Also changed the way errors are passed to the client and
parsed on client side from simple string to JSON containing much
more details and old school message inside.

Conflicts:
	src/api-lib/vnc_api.py
	src/config/api-server/vnc_cfg_api_server.py

Closes-Bug: #1708697
Change-Id: I78a6883d228b38cc28c830aa63fe4df1ea55410a