RBAC: contrail-api exceptions during stack creation/update
Description: cassandra object memory cache had a bug, where
the time stamp of an object was not updated, when ref in
that object is created/deleted. This was resulting in
having stale cached object in a api-server while while the
operation happened in another api-server. Fixed it.
Another issue, while working on the cached refs to obtain
perms2 associated with the ref objects, it is possible
that some ref objects might have got deleted in another
api-server context. Links should be considered for the available
refs only. Fixed it. Both the issues are not related to RBAC,
it is just that these race conditions are exposed on enabling
RBAC.

Change-Id: I916cbb4e653192d06f4b8135bd3e7d3a398e07cf
Closes-Bug:#1706007