Modify class VncAmqpHandle constructor
VncAmqpHandle takes service's config as an argument (args) in the constructor
and extracts rabbitmq parameters and other config parameters like
trace_file from the args. We plan to use VncAmqpHandle in analytics
services and the rabbitmq parameter names are not the same as the config
service. So, instead of passing args to VncAmqpHandle, pass the rabbitmq
config as dict.

Change-Id: I6a9be6e3aa71af58342857c443beb8f211b80816
Related-Bug: #1703709