Fix PRouter UVE name sent by snmp-collector
If the user configures a physical router with the name different from
the hostname of the physical router, then snmp-collector sends PRouter
UVE entry with hostname (snmp name) of physical router instead of the
name configured by the user. Fix this by sending the PRouter UVE with
user configured name.

Change-Id: I2b3f19b0f180b95faabe4f9cd5db72de25de8717
Closes-Bug: #1708727