Must call zkclient.close after every connect attempt
Zookeeper client opens a socketpair and sends a single byte message.
There is no receive on this socket until it is connected successfully.
If it never connects, we will eventually run out of socket buffer space.
For this reason, we should call close() to close these sockets after
each connection attempt.

Change-Id: Iefdc8481d07bf243d70fb8f7e22792f4c65a6cbf
Closes-Bug: 1691541