portbinding values must be strings, not dict
We store portbinding value as string (either plain or json encoded).
If this is stored as dict, the client will not be able to parse it correctly.
We made this change earlier, but one of the bindings that is created internally
is still being set as string. With this code, it will be converted to json
string before writing to database.

Change-Id: I0aa0040bb76c8f4eecb7530c0a744a5bc31b5daf
Closes-Bug: 1709822