Fixed service delete issue in k8s-nested
In the service delete, floating-ips for lb's vmi would
be deleted, before the loadbalancer delete is called.
In the loadbalancer delete also the floating-ips would
be deleted while deleting the vmi. Since it is already
deleted and it not updated in the local-cache, NoIdError
exception is got which is handled.

Change-Id: Ia615e74cce588c948f94c35406a3b0c1cb3c357f
Closes-bug: #1695219