Handle null lldpRemSysName in lldpTable
In MX, if lldp is enabled on the interface which is connected to compute node,
then the lldp mib table has null value for lldpRemSysName and lldpRemPortDesc.
The existing logic doesn't consult mac or arp table if the entry is present in
the lldp table. Therefore, if the lldpRemSysName or lldpRemPortDesc is null in
the lldpTable, ignore this entry to proceed with fdb/arp table lookup.

Change-Id: Iabc285730381b0032c305186b8c7bbfb759c8c4a
Closes-Bug: #1708758