RBAC: Multiple RBAC issues handled as part of this commit
Description: cassandra object memory cache had a bug, where
the time stamp of an object was not updated, when ref in
that object is created/deleted. This was resulting in
having stale cached object in a api-server while while the
operation happened in another api-server. Fixed it.
Another issue, while working on the cached refs to obtain
perms2 associated with the ref objects, it is possible
that some ref objects might have got deleted in another
api-server context. Links should be considered for the available
refs only. Fixed it. Both the issues are not related to RBAC,
it is just that these race conditions are exposed on enabling
RBAC.

Closes-Bug:#1706007

Description: Multiple issues are seen here.
In the API server, project id was not in uuid form.
Sometimes a project may not have an associated domain.
This will make shared=true work on API server.
Neutron relies on API lib to post the requests, so made
change in library to accept information and is backward
compatible for other applications.
The behavior in neutron going to be as follows:
1) neutron net-list would list
   a) all local networks in that tenant.
   b) globally shared.
   c) tenant shared.
2) neutron net-list --shared True would list
   a) globally shared.
   b) tenant shared.
3) neutron net-list --shared False would list
   a) all local networks in that tenant.

Closes-Bug:1707175

Description: This script helps to take snapshot and restore it
on local environment. Very useful in testing and debugging customer
issues. Instruction on how to run are available in that script
itself.

Change-Id: I84046f2fab42060a99b47d873b7f9a74c399e208
Closes-Bug:#1709498