Making VNC API pip installable
VNC API is much dependent on the config common code right now. Also we are
hard coding the version of the generated package. Which makes the package
static and dependent on other code. Hence, to upload to pip as a package,
we need to make it independent of the other modules and also automate the
versioning of the package.

Fixed all pep8 errors in vnc_api module

Change-Id: Idc914c445d9ef18d4c2e2313298bf2c4075375fc
Partial-Bug: #1711447