Making VNC API pip installable
VNC API is much dependent on the config common code right now. Also we are
hard coding the version of the generated package. Which makes the package
static and dependent on other code. Hence, to upload to pip as a package,
we need to make it independent of the other modules and also automate the
versioning of the package.

Change-Id: I93eaa3e622613cd2f8b9393184301c155d85e6a2
Partial-Bug: #1711447