Operations on objects created pre RBAC fail, on enabling RBAC.
Description: Certain objects created pre RBAC has owner ID not
same as the id of the project in which those objects are created.
This is because the project information passed was the service
tenant. This is causing failure of operations on those objects
after enbaling RBAC.
This change will set appropriate tenant ids as the owner.

Change-Id: I108a7543b5d1241d85471382bb3779823e86b0dd
Closes-Bug: #1706218