Fixed Traceback issues in svc-monitor->instance_manager
1. Fixed NoIdError Traceback
During vmi delete, floating-ip update would be done
which will try to do the ref-update between floating-ip
and vmi. Since it is a vmi delete operation, sometimes
vmi may not be there when trying to do the ref-update
which caued the exception. Added the NoIdError Exception
to handle this.
2. Fixed BadRequest Traceback
If the si.parent_key is None, It may be happen.
Using the proj_obj.uuid would avoid the issue.

Change-Id: If11e09ddd5c9028af68af95e016c9d35a58d0064
Partial-bug: #1641364