Increase the task monitor timeout
In the crash seen in this bug, task monitor asserts although the task
activity seems to be going fine. Agent in not in gdb, no system time
change is seen (even if server is healvily loaded this should not be
seen). As no other task monitor bug is seen, increasing the timeout
to 20s.

Change-Id: Iceb277c2b25c3041175c158efa3cb1c4bbbf91d3
partial-bug: #1714744