Fixed service-instance delete issue
During the loadbalancer delete, If the instance ip
is removed before the service instance, the evalute of
service instance would call the create_service_instance()
which would do the chown for the instance_ip which may not
be there which would cause this error.
Added exception handler to handle this.

Change-Id: I7688b8818f4b68c117ebc024d192fa3f2b4454fb
Closes-bug: #1716881