Make corefile_path variable adjustable.
When the first character of /proc/sys/kernel/core_pattern file
is a | symbol, the remainder of the line is to be interpreted
as a command-line to a userspace script that will process the
coredump file. In that case we cannot parse that line and assume
it will point to the location of system core files. Instead, allow
operators to specify where core files are located.

Closes-Bug: #1717862
Change-Id: Ibbde444fdf555dccba41c8c9e26846f2570e4e7b