Explicitly define RapidJSON products
Define a product of RapidJSON build (just headers) to be consumed by
targets that depend on them.

Remove explicit dependency on '#build/include/rapidjson/', as depending
on directories is not officially supported by Scons and the
OpServerProxy.cc dependencies should be handled by Scons anyway now that
it knows where to look for them.

Partial-Bug: #1476261
Change-Id: Ic59aa3f378bf50e5f4c2d4f9388374a0d4c116a6