Invalid return value while capturing auth hdrs
Description: Even on token expiry, registered function
is called to extract headers, and we return None. The middleware
expects empty dict.
But, in neutron plugin for failed authentication we need to raise
401 exception.
To keep UT happy, rbac would return 401 exception when roles are
missing instead of 403, and move out the logic from neutron plugin.

For some reason, api-server through neutron is not sending right
exception to neutron, which bottle treats it as 500. That will be
handled as second commit. So this commit is partial bug, and also
validate the change against systest scripts.

Change-Id: Ie05a6046e52d7ab3bd04cb7dc5fb8c0c3cbf2b34
Partial-Bug: #1718056