stop/start/restart a specific contrail-alarm-gen instance
When there are multiple instances of contrail-alarm-gen services running
on the same server, user should be able to stop/start/restart a specific
contrail-alarm-gen instance.
For example, if there are 2 instances of contrail-alarm-gen service,
then the name of the instances would be contrail-alarm-gen:0 and
contrail-alarm-gen:1
The following command can be used to restart contrail-alarm-gen:1
"service contrail-alarm-gen 1 restart"; where 1 is the instance id.
If the instance id is not specified, then all the contrail-alarm-gen
instances would be restarted.

Change-Id: Ia580f19b3ddf850c586da302d3db5db7eaebc5bd
Partial-Bug: #1714000