floating ip ownership is incorrectly set for calls from neutron
Description: Floating IP ownership has to be explictly set, to
the project in which it is being created. Otherwise reads will
fail.
Flating IPs are created on shared networks or shared floating
ip pools. The ownership of the floating ip should be in the
tenant in which it was created and not from the project in the
fqname, which is default ownership setting.
So, for this, we need to set the owner in the floatingip
object, while creating it.

Change-Id: Ia27eca5dd81658fead072e1fdd9338cbbbacf040
Closes-Bug: #1720473