[VNC config] Default APS for all applications
Add a scoped default Application Policy Set for all applications.
It creates/deletes that APS for all projects with the name
'default-application-policy-set' and the flag 'all_applications' set to
True. That new flag 'all_applications' replaces the old one 'is_global',
it is only readable and uses for the default APSs.
The global default application set is retained but renamed
'default-application-policy-set' and use the 'all_application' flag and
stays a child of the global Policy Management resource.

Closes-Bug: #1718012
Change-Id: I9f53c0f21983bf191b4c51318745eb348d48dd86