Contrail RBAC - field option does not support multilevel fields rbac rule aggregation not working correctly
Description: Rules configured against multiple access-lists at domain
global-system-config or project are not aggregated.
This will aggregate rules.

Added support for multilevel field rules for
with create and update operations. For delete and read, individual
fields are not present in the dict. So, rules are no op for those
operations.

Change-Id: I76b0db0630ea8ecd75365b4553db146717b62343
Closes-Bug: #1719040
Closes-Bug: #1702321
Closes-Bug: #1699099