Return empty list if obj_uuid, back_ref_id, fip_objs are empty
1706260 fix is need for 1718972.

Description of 1718972:
fip_objs are returned empty, return empty list.

Description of 1706260:
If a floating ip is not attached to a port, in vnc_openstack code,
we might call virtual_machine_interfaces_list with obj_uuids set to
empty. In this case, current library code will not set that field
in the request body causing API server to return all VMIs. To avoid
this situation, we now return early with empty list if either obj_uuids
or back_ref_id is set as empty list. Old behavior will remain if any
of these arguments are not passed or are passed as None.
Backporting to older releases.

Change-Id: Ib7453cd71952288bc91542515d5381b935292705
Closes-Bug: 1706260
Closes-Bug: 1718972