Update source address while sending icmpv6 error packet
When source VM and service instance are on the same compute, ICMP error for
too big packet will have to be sent for packet coming out of the right interface
of the service instance. Agent was trying to find the associated IPAM for the
interface using the original address, which is failing resulting in ICMP not
being sent. Change to use vhost address mapped IPv6 address as the source
to avoid this.

Change-Id: I1f137afba59edee4cf86b423caf25c472d55f365
closes-bug: #1721457
(cherry picked from commit ccbca452d4950790a0c0a918bc41568df6d234d7)