Fix tsn agent crash
In TSN mode, ignore updation of flow-based statistics as they
never get exported.

The crash was seen because structures required for updation of
flow-based statistics are not initialized in TSN mode and we were
trying to update the statistics on them.

Closes-Bug: #1719578
(cherry picked from commit 70a8e018379757ca43572cf6931a9b2cf7666602)

Change-Id: Id12dba5564f39a3e2e58e75bb9f23a3ba22f8220