Fix tsn agent crash
In TSN mode, ignore updation of flow-based statistics as they
never get exported.

The crash was seen because structures required for updation of
flow-based statistics are not initialized in TSN mode and we were
trying to update the statistics on them.

Closes-Bug: #1719578
(cherry picked from commit 70a8e018379757ca43572cf6931a9b2cf7666602)

Change-Id: I185f2683bd26cb005252144d133a73911ed89039