[VNC OpenStack] Build router GW VMI directly
The Neutron router gateway creation is asynchronous. First user set the
external gateway to the router, that creates a link between LR and
VN. Then svc-monitor SNAT agent reacts on that link creation and
instantiates a SNAT SI. Finally, svc-monitor creates SI resources (VM, VMI
and IIP) and schedules VMs on vrouters.
Until all that stuff is done, the Neutron router does not have any port
corresponding to the router external gateway interface. That patch adds
new attribute to the LR to VMI reference to distinguish internal router
interfaces and external gateway interface. And when a gateway is set,
vnc_openstack code creates the gateway VMI and references it from the
LR. Svc-monitor will re-use that VMI later when it will build SI.

Change-Id: I74d0a87a392f8825c20769388d2fa92cb0139964
Closes-Bug: #1725370