[VNC OpenStack] Fix concurrency project deletion
In mutli API server setup, the sync domains/projects VNC OpenStack
greenthread of one of the servers can try to remove and purge resource related
to a project which another server already purged. That results to not found
API exceptions. That patch catches that exceptions and abandons to purge
the project.
That also reveals an issue in permission code that checks if the
deletion is authorized which fails to read the resource as it was
already deleted.

Change-Id: I140cf5b7f4bd4d7798db18a7afd3b0d6168ea6b1
Closes-Bug: #1724179
Closes-Bug: #1724194