Set DHCP siaddr to IP of the TFTP server if set
DHCP clarifies the interpretation of the 'siaddr' field as the
address of the server to use in the next step of the client's
bootstrap process according to rfc 2131
https://www.ietf.org/rfc/rfc2131.txt

TSN node doesn't provide tftp service, and for baremetal provisioning
the tftp is located on ironic-conductor host.

This patch update DHCP to set siaddr to value set in option 66 if
provided.

Closes-Bug: #1724595

Change-Id: I1da95e7d06f8a2390556c168dac8e33a30aea7a7