misc fixes in RBAC, descriptive exception messages and floatingip-list latency
Description for 1723278: Forcing is_shared and perms2.global_access to be
consistent on restart of api-server.

Descrption for 1721609:
1) Domain children objects, should have owner as tenant from env
in which it is created.
2) More descriptive error messages. More changes are need in neutron
plugin code to relay the error message from api-server to neutron
client.
3) Missing exception handling code.
4) Latency issue for floating IP list from neutron.
5) uuid and fq_name fields are filtered from subjecting to rbac rules.
 and its unit test.

Description for 1719040 and 1702321:
Rules configured against multiple access-lists
at domain global-system-config or project are not aggregated.
This will aggregate rules.
rbac rule aggregation not working correctly

Description of 1718972:
fip_objs are returned empty, return empty list.

Change-Id: I21ba51c155d746b634edfc3533c4aba3c0c2442e
Closes-Bug: #1702321
Closes-Bug: #1719040
Closes-Bug: #1721609
Closes-Bug: #1723278
Closes-Bug: #1718972