floating ip ownership is incorrectly set for calls from neutron
Description: Floating IP ownership has to be explictly set, to
the project in which it is being created. Otherwise reads will
fail.
Flating IPs are created on shared networks or shared floating
ip pools. The ownership of the floating ip should be in the
tenant in which it was created and not from the project in the
fqname, which is default ownership setting.
So, for this, we need to set the owner in the floatingip
object, while creating it.

UT for 1720473

Change-Id: I8507cb90883c0543b341a9bb2e9c1d753d43f483
Closes-Bug: #1720473
Closes-Bug: #1721445