Publish disk_usage and compaction_tasks in redis
As part of https://bugs.launchpad.net/juniperopenstack/+bug/1604200,
feedback loop was created between contrail-analytics-api and
contrail-collector to throttle/stop writing data in cassandra based on
the disk usage and # compaction tasks. But the commit for #1677694
accidentally removed the code that sets the watermark in
contrail-collector.
Added a new task "monitor_analytics_db" to periodically read the
database-node UVE and publish the disk_usage_percentage and
pending_compaction_tasks in redis.

Change-Id: I30ae68a13ec216d55d6bc7efd9e834b8da74c910
Closes-Bug: #1724659