Support VM interface to bgp-router link.
As mentioned in the bug under
https://bugs.launchpad.net/juniperopenstack/+bug/1725777/comments/1 multiple VMI
will be linked to bgp-as-service object which in turn is linked to many bgp
routers. This used to create confusion in agent to select router among list of
routers to connect to VMI. Agent will connect to first one available.
Introduction of this link will provide guidance to agent to find bgp-router to
be used for every VMI.

Partial-bug: #1725777
Change-Id: Id030f6c3dc47e13e8d9640b0a9efc1041e87ec22